+++ GPTifier/CMakeLists.txt
@@ -80,7 +80,6 @@ add_executable(
   "gpt"
   "src/cli.cpp"
   "src/command_chats.cpp"
-  "src/command_files.cpp"
   "src/command_fine_tune.cpp"
   "src/command_test.cpp"
   "src/configs.cpp"
@@ -88,6 +87,7 @@ add_executable(
   "src/help_messages.cpp"
   "src/interface/command_costs.cpp"
   "src/interface/command_embed.cpp"
+  "src/interface/command_files.cpp"
   "src/interface/command_models.cpp"
   "src/interface/command_run.cpp"
   "src/interface/command_short.cpp"
